backend/requirements.txt
fastapi
pillow
uvicorn
python-multipart
jax[cpu]
dm-haiku
optax
numpy
matplotlib
